requirements.txt
SQLAlchemy==1.3.22
Flask_JWT_Extended==4.1.0
flask_marshmallow==0.14.0
pytest==6.2.2
facenet_pytorch==2.5.1
flask_restx==0.2.0
passlib==1.7.4
Werkzeug==1.0.1
Flask_SQLAlchemy==2.4.4
torch==1.7.0
numpy==1.19.2
torchvision==0.8.1
Flask==1.1.2
Pillow==8.2.0
